+++ pages/transection-flow-orders.js
@@ -5,9 +5,7 @@ import styles from '@/styles/Home.module.css'
 import { useEffect, useState } from 'react'
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/router'
-import CustomModal from '../components/CustomModal';
-import { toast } from 'react-toastify';
-
+import DataTable from '../components/DataTableTransactionFlow';
 
 const inter = Inter({ subsets: ['latin'] })
 import axios from 'axios';
@@ -16,8 +14,6 @@ export default function Product() {
   
 
   const [data, setData] = useState([]);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [datas, setDatas] = useState(null);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -31,32 +27,6 @@ export default function Product() {
 
     fetchData();
   }, []);
-
-  const closeModal = () => {
-    // Close the modal and clear the data
-    setIsModalOpen(false);
-    setDatas(null);
-  };
-
-
-  const handleButtonClick = async (data) => {
-    
-    try {
-      // Make API request with the provided id
-      const response = await fetch(`https://bitapi.mfhsoltech.com/api/orders/${data.id}`);
-      const result = await response.json();
-
-      
-      setDatas(result);
-      // Open the modal
-      setIsModalOpen(true);
-      
-    } catch (error) {
-      toast.error('Something went Wrong');
-      console.error('Error fetching data:', error);
-    }
-  };
-
 
 
   return (
@@ -81,63 +51,8 @@ export default function Product() {
                       </h4>
                       
                     </div>
-                    <div className="card-body px-0 py-1">
-                    {data.length > 0 ? (
-                    <table className="table table-responsive">
-                      <thead>
-                        <tr>
-                                          
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>order number</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Member ID</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>username</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>order time</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>product information</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>state</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>direction</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>time/points</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Position opening point</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>closing point</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Commission balance</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Invalid order balance</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Effective order balance</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Actual profit and loss</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Balance after purchase</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Single control operation</th>
-                        <th scope="col" style={{ whiteSpace: 'nowrap' }}>Details</th>
-                          
-                        </tr>
-                      </thead>
-                      <tbody>
-                      {data.map((record) => (
-                        <tr key={record.id}>
-                          <td>{record.id}</td>
-                          <td>{record.MemberId}</td>
-                          <td>{record.username}</td>
-                          <td>{record.ordertime}</td>
-                          <td>{record.productInfo}</td>
-                          <td>{record.state}</td>
-                          <td>{record.direction}</td>
-                          <td>{record.timePoints}</td>
-                          <td>{record.positionOpenPoint}</td>
-                          <td>{record.closingPoint}</td>
-                          <td>{record.commissionBalance}</td>
-                          <td>{record.invalidOrderBalance}</td>
-                          <td>{record.effectiveOrderBalance}</td>
-                          <td>{record.actualProfitAndLoss}</td>
-                          <td>{record.balanceAfterPurchase}</td>
-                          <td> {record.singleControlOperation}</td>
-                          <td>
-                            <button className='btn btn-success mb-1' onClick={() => handleButtonClick({id: record.id })}>Check</button>
-                            <button className='btn btn-danger'>Trash</button>
-                          </td>
-                        </tr>
-                         ))}
-                      </tbody>
-                     
-                    </table>   
-                       ) : (
-                        <p>Loading...</p>
-                     )} 
+                    <div className="card-body px-0 py-1 p-3">
+                       <DataTable data={data}/>
                     </div>
                    <div>
                    </div>
@@ -151,8 +66,6 @@ export default function Product() {
       </div>
     </div>
   </div>
- 
-  <CustomModal isOpen={isModalOpen} onClose={closeModal} data={datas} /> 
 
 </div> 
  );
