+++ pages/user-profile.js
@@ -4,26 +4,55 @@ import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 import { useEffect, useState } from 'react'
 import { useSession } from 'next-auth/react'
-import { useRouter } from 'next/router'
 import { toast } from 'react-toastify';
 import axios from 'axios';
 
 
 const inter = Inter({ subsets: ['latin'] })
 
-export default function Product() {
+export default function Profile() {
   
 
-  const [data, setData] = useState([]);
-  const router = useRouter();
+  
+  const [error, setError] = useState(null);
+  const [newName, setNewName] = useState('');
+  const [newNameEmail, setEmailName] = useState('');
+  
+
 
   useEffect(() => {
     const fetchData = async () => {
-      try {  
-        const response = await axios.get('https://bitapi.mfhsoltech.com/api/withdrawal/index');
-        setData(response.data.withdrawal_records);
+      try {
+
+
+        const accessToken = localStorage.getItem('token');
+
+        const response = await fetch('https://bitapi.mfhsoltech.com/api/profile', {
+          method: 'GET',
+          headers: {
+            'Authorization': `Bearer ${accessToken}`,
+            'Content-Type': 'application/json',
+          },
+        });
+
+        const data = await response.json();
+
+        if (response.status === 200) {
+          setNewName(data.details.name)
+          
+          
+        } else if (response.status === 401) {
+          
+          toast.error('Login Session Expired'); 
+        } else {
+         
+          toast.error('Something Went wrong');
+          
+        }
       } catch (error) {
-        console.error('Error fetching data:', error);
+        
+          toast.error('Something Went wrong');
+          
       }
     };
 
@@ -31,9 +60,61 @@ export default function Product() {
   }, []);
 
 
-  const handleButtonClick = (id) => {
+
+  const handleNameChange = (e) => {
+    setNewName(e.target.value);
+  };
+
+  const handleEmailChange = (e) => {
+    setEmailName(e.target.value);
+  };
+
+
+
+  const updateProfile = async (username, password) => {
     
-    router.push(`/withdrawal-details?id=${id}`);
+        const accessToken = localStorage.getItem('token');
+
+    const apiUrl = 'https://bitapi.mfhsoltech.com/api/profile';
+  
+    try {
+      // Make a POST request using the fetch function
+      const response = await fetch(apiUrl, {
+        method: 'POST',
+        headers: {
+          'Authorization': `Bearer ${accessToken}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          name: username,
+          password: password,
+        }),
+      });
+  
+      // Check if the request was successful (status code 2xx)
+      if (response.status === 200) {
+        const data = await response.json();
+
+        toast.success('Profile Updated Successfully'); 
+
+      } else if (response.status === 401) {
+
+        toast.error('Login Session Expired'); 
+
+      } else {
+
+        toast.error('Something Went wrong');
+
+      }
+
+    } catch (error) {
+      
+      console.error('Error during API request:', error);
+    }
+  };
+  
+  const handleUpdateProfile = () => {
+    updateProfile(newName, newNameEmail);
   };
 
 
@@ -55,54 +136,39 @@ export default function Product() {
                   <div className="card widget-todo">
                     <div className="card-header border-bottom d-flex justify-content-between align-items-center flex-wrap">
                       <h4 className="card-title d-flex mb-25 mb-sm-0">
-                        <i className="bx bx-check font-medium-5 pl-25 pr-75" /> Record WithDrawal
+                        <i className="bx bx-check font-medium-5 pl-25 pr-75" /> Profile
                       </h4>
                       
                     </div>
                     <div className="card-body px-0 py-1">
-                    {data.length > 0 ? (
-                    <table className="table table-responsive">
-                      <thead>
-                        <tr>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>serial number</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Account</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Name</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Withdrawal amount (yuan)</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>handling fee</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Actual arrival (yuan)</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Bank of deposit</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Denial Reason</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Submission time</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Processing progress</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>operate/Action</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                      {data.map((record) => (
-                        <tr key={record.id}>
-                            <td>{record.serialNum}</td>
-                            <td>{record.MemberId}</td>
-                            <td>{record.username}</td>
-                            <td>{record.withdrawalAmount}</td>
-                            <td>{record.handlingFee}</td>
-                            <td>{record.actualArrival}</td>
-                            <td>{record.bankDeposit}</td>
-                            <td>{record.denialReason}</td>
-                            <td>{record.created_at}</td>
-                            <td>{record.processingProgress}</td>
-                            <td>
-                              <button className='btn btn-success mb-1' onClick={() => handleButtonClick(record.id)}>Click me Details</button>
-                              <button className='btn btn-danger'><i className='fa fa-trash'></i></button>
-                            </td>
+                    
+                   <form  className='p-3'>
+                    <div className='container'>
+                      <div className='row'>
+                       
+                        <div className='col-md-6'>
+                            <label>Name:  </label>
+                            
+                            <input type="text" className='form-control' value={newName} onChange={(e) => setNewName(e.target.value)} />
 
-                        </tr>
-                         ))}
-                      </tbody>
-                     
-                    </table>   
-                       ) : (
-                        <p>Loading...</p>
-                     )} 
+                        </div>
+                        
+                        <div className='col-md-6'>
+                            <label>Password:  </label>
+                            <input type="password" className='form-control' value={newNameEmail} onChange={(e) => setEmailName(e.target.value)} />
+
+                        </div>
+                        <div className='col-md-12 mt-3'>
+                            
+                        <button type="button" className='form-control btn btn-success' onClick={handleUpdateProfile}>Update Prfile</button>
+
+                        </div>
+
+                      </div>
+                     </div> 
+                    
+                  </form>
+                  
                     </div>
                    <div>
                    </div>
