+++ pages/withdrawal-record-user-center.js
@@ -4,9 +4,9 @@ import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 import { useEffect, useState } from 'react'
 import { useSession } from 'next-auth/react'
-import { useRouter } from 'next/router'
 import { toast } from 'react-toastify';
 import axios from 'axios';
+import DataTable from '../components/DataTableWithdrawalRecord';
 
 
 const inter = Inter({ subsets: ['latin'] })
@@ -15,7 +15,7 @@ export default function Product() {
   
 
   const [data, setData] = useState([]);
-  const router = useRouter();
+
 
   useEffect(() => {
     const fetchData = async () => {
@@ -30,11 +30,6 @@ export default function Product() {
     fetchData();
   }, []);
 
-
-  const handleButtonClick = (id) => {
-    
-    router.push(`/withdrawal-details?id=${id}`);
-  };
 
 
   return (
@@ -59,50 +54,11 @@ export default function Product() {
                       </h4>
                       
                     </div>
-                    <div className="card-body px-0 py-1">
-                    {data.length > 0 ? (
-                    <table className="table table-responsive">
-                      <thead>
-                        <tr>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>serial number</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Account</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Name</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Withdrawal amount (yuan)</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>handling fee</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Actual arrival (yuan)</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Bank of deposit</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Denial Reason</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Submission time</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>Processing progress</th>
-                          <th scope="col" style={{ whiteSpace: 'nowrap' }}>operate/Action</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                      {data.map((record) => (
-                        <tr key={record.id}>
-                            <td>{record.serialNum}</td>
-                            <td>{record.MemberId}</td>
-                            <td>{record.username}</td>
-                            <td>{record.withdrawalAmount}</td>
-                            <td>{record.handlingFee}</td>
-                            <td>{record.actualArrival}</td>
-                            <td>{record.bankDeposit}</td>
-                            <td>{record.denialReason}</td>
-                            <td>{record.created_at}</td>
-                            <td>{record.processingProgress}</td>
-                            <td>
-                              <button className='btn btn-success mb-1' onClick={() => handleButtonClick(record.id)}>Click me Details</button>
-                              <button className='btn btn-danger'><i className='fa fa-trash'></i></button>
-                            </td>
+                    <div className="card-body px-0 py-1 p-3">
+                 
+                       <DataTable data={data}/>
 
-                        </tr>
-                         ))}
-                      </tbody>
                      
-                    </table>   
-                       ) : (
-                        <p>Loading...</p>
-                     )} 
                     </div>
                    <div>
                    </div>
